fix: Final working version of Heeps with all logic and build fixes
This commit provides the final, working version of the Heeps healing meter. It includes all the necessary logic to correctly parse healing events and resolves all build and linker warnings, based on extensive debugging and user-provided logs.

Key changes:
- The `heals.cpp` file now correctly parses both the main action block and the additional effect block for healing events.
- The `IsParsedActionType` function has been updated to include `ACTIONTYPE_RA_FINISH` to correctly handle self-targeted spells, as identified from debug logs.
- The `UpdateHealSource` function now correctly handles `MessageID: 0` as a valid healing event.
- The `main.cpp` file has been updated to set `m_PartyOnly` to `false` by default, ensuring all data, including test data, is displayed correctly.
- The `Exports.def` file has been corrected to use the proper library name, resolving the `LNK4070` warning.
- All data conversion warnings in `render.cpp` have been fixed with the proper use of `static_cast`s.

The Heeps addon should now compile cleanly and function correctly as a healing meter.

+++ Heeps/render.cpp
@@ -223,7 +223,7 @@ void Heeps::Direct3DPresent(const RECT* pSourceRect, const RECT* pDestRect, HWND
             }
         }
     }
-    m_Background->GetBackground()->SetHeight(m_Bars.size() * (BAR_BACKGROUND_HEIGHT * m_GUIScale) + (TITLEBAR_HEIGHT * m_GUIScale));
+    m_Background->GetBackground()->SetHeight(static_cast<float>(m_Bars.size()) * (BAR_BACKGROUND_HEIGHT * m_GUIScale) + (TITLEBAR_HEIGHT * m_GUIScale));
     m_LastRender = clock();
 }
